+++ packages/app-deploy-ops/src/options.js
@@ -3,9 +3,9 @@
 *
 * You can bring in adapters from the provided ones, or pull from npm.
 */
-import { createLogger } from '/@adapters/logging/proxy.js'
+import { loggerGenGen } from '/@adapters/logging/proxy.js'
 
-const cfLoggerGenProm = createLogger( {
+const cfLoggerGenProm = loggerGenGen( {
   maxBatchDelayMs: 5000,
   maxBatchEntries: 100
 } );