+++ packages/app/cypress/support/index.js
@@ -1,8 +1,0 @@
-//
-// cypress/support/index.js
-//
-// Collects together custom Cypress commands.
-//
-import './firebase'
-import './auth'
-