+++ packages/app/cypress/users.js
@@ -1,21 +1,21 @@
 /*
-* cypress/users.js
+* cypress/users/index.js
 *
-* Users used in tests (their uid's are in the test data).
-*
-* Note: You can also log in with these, in 'npm run dev:test'.    <-- tbd. nope
+* Users used in tests (cypress/integration).
 */
 const users = [
   {
     uid: "joe",
-    name: "Joe D.",
+    displayName: "Joe D.",
     photoURL: "https://i.pinimg.com/originals/2b/59/6d/2b596d481a2296297d8098e09bfd44e5.jpg"
   },
   {
     uid: "william",
-    name: "William Dalton",
+    displayName: "William Dalton",
     photoURL: "https://pbs.twimg.com/profile_images/508788470479720449/GagopTx-_400x400.png"
   }
 ]
 
-export { users }
+const joe = users.find( x => x.uid === 'joe' );
+
+export { users, joe }